Manage map in context provider
